Move everything to React
