+++ app/components/Issue.js
@@ -2,7 +2,7 @@
 
 import React, { Component } from 'react'
 import Blob from './Blob'
-import c from './constants'
+import c from '../constants'
 
 export default class Issue extends Component {
   render () {
@@ -14,7 +14,7 @@ export default class Issue extends Component {
       )
     }
 
-    const Footer = () => [<hr key={1} />, <h6 key={2}>This issue was generated by <a href={c.app}>todo</a> based on a <code>{c.cfg.keyword}</code> comment in <a href="#"><code>{c.cfg.sha}</code></a>. It's been assigned to <a href="https://github.com/JasonEtco" className="user-mention">{c.cfg.author}</a> because they committed the code.</h6>]
+    const Footer = () => [<hr key={1} />, <h6 key={2}>This issue was generated by <a href={c.app}>todo</a> based on a <code>{c.cfg.keyword}</code> comment in <a href="#"><code>{c.sha}</code></a>. It's been assigned to <a href="https://github.com/JasonEtco" className="user-mention">{c.cfg.author}</a> because they committed the code.</h6>]
 
     return (
       <div className="issue rounded-1">
@@ -55,7 +55,7 @@ export default class Issue extends Component {
                     <td className="d-block comment-body markdown-body">
                       {body && <p>{body}</p>}
                       {body && <hr />}
-                      <Blob start={start} code={code} sha={c.cfg.sha} />
+                      <Blob start={start} code={code} sha={c.sha} />
                       <Footer />
                     </td>
                   </tr>
