+++ app/index.tpl.html
@@ -5,26 +5,16 @@
     <title>todo | GitHub App</title>
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <link href="https://fonts.googleapis.com/css?family=Roboto" rel="stylesheet">
-    <meta name="theme-color" content="#fe6300">
+    <meta name="theme-color" content="#00B0D8">
 </head>
-<body class="Site">
+<body>
     <noscript>
         <div class="no-script">
             <h1>You must enable JavaScript to use try out todo.</h1>
             <a href="https://github.com/JasonEtco/todo/issues" target="_blank" rel="noopener noreferrer">Need help?</a>
         </div>
     </noscript>
-    <header class="Site-header bg-gray-dark py-2 text-white text-center">
-        <h1 class="d-flex flex-items-center flex-justify-center"><img class="rounded-1 mr-2" src="https://avatars1.githubusercontent.com/in/5534?v=4&s=32" /> todo</h1>
-        <a href="https://github.com/JasonEtco/todo" class="text-white text-bold">Contribute on <svg class="octicon octicon-mark-github" viewBox="0 0 16 16" version="1.1" width="16" height="16" aria-hidden="true"><path fill-rule="evenodd" d="M8 0C3.58 0 0 3.58 0 8c0 3.54 2.29 6.53 5.47 7.59.4.07.55-.17.55-.38 0-.19-.01-.82-.01-1.49-2.01.37-2.53-.49-2.69-.94-.09-.23-.48-.94-.82-1.13-.28-.15-.68-.52-.01-.53.63-.01 1.08.58 1.23.82.72 1.21 1.87.87 2.33.66.07-.52.28-.87.51-1.07-1.78-.2-3.64-.89-3.64-3.95 0-.87.31-1.59.82-2.15-.08-.2-.36-1.02.08-2.12 0 0 .67-.21 2.2.82.64-.18 1.32-.27 2-.27.68 0 1.36.09 2 .27 1.53-1.04 2.2-.82 2.2-.82.44 1.1.16 1.92.08 2.12.51.56.82 1.27.82 2.15 0 3.07-1.87 3.75-3.65 3.95.29.25.54.73.54 1.48 0 1.07-.01 1.93-.01 2.2 0 .21.15.46.55.38A8.013 8.013 0 0 0 16 8c0-4.42-3.58-8-8-8z"></path></svg></a>
-    </header>
-
-    <div class="Site-content mount pt-2 px-2"></div>
-
-    <footer class="Site-footer mt-6 py-4 bg-gray-light border-top text-center">
-        <p class="m-0">You can install <strong>todo</strong> in your repositories <a href="https://github.com/apps/todo">right here.</a></p>
-        <p class="mt-0 mb-1">Built by <a href="https://twitter.com/JasonEtco">@JasonEtco</a> with 🤖 using <a href="https://github.com/probot/probot">Probot</a>.</p>
-    </footer>
+    <div class="mount Site"></div>
     <script src="https://cdn.polyfill.io/v2/polyfill.min.js"></script>
 </body>
 </html>
