+++ app/style.scss
@@ -17,7 +17,7 @@
   height: auto;
 }
 
-html { height: 100% }
+html, body { height: 100% }
 
 .Site {
   display: flex;
